+++ packages/google_maps_flutter/google_maps_flutter/lib/google_maps_flutter.dart
@@ -30,6 +30,8 @@ export 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
         Cluster,
         ClusterManager,
         ClusterManagerId,
+        GroundOverlay,
+        GroundOverlayId,
         Heatmap,
         HeatmapGradient,
         HeatmapGradientColor,