+++ packages/google_maps_flutter/google_maps_flutter/lib/src/controller.dart
@@ -43,8 +43,9 @@ class GoogleMapController {
           .listen((_) => _googleMapState.widget.onCameraMoveStarted!());
     }
     if (_googleMapState.widget.onCameraMove != null) {
-      GoogleMapsFlutterPlatform.instance.onCameraMove(mapId: mapId).listen(
-          (CameraMoveEvent e) => _googleMapState.widget.onCameraMove!(e.value));
+      GoogleMapsFlutterPlatform.instance
+          .onCameraMove(mapId: mapId)
+          .listen((CameraMoveEvent e) => _googleMapState.widget.onCameraMove!(e.value));
     }
     if (_googleMapState.widget.onCameraIdle != null) {
       GoogleMapsFlutterPlatform.instance
@@ -54,17 +55,18 @@ class GoogleMapController {
     GoogleMapsFlutterPlatform.instance
         .onMarkerTap(mapId: mapId)
         .listen((MarkerTapEvent e) => _googleMapState.onMarkerTap(e.value));
-    GoogleMapsFlutterPlatform.instance.onMarkerDragStart(mapId: mapId).listen(
-        (MarkerDragStartEvent e) =>
-            _googleMapState.onMarkerDragStart(e.value, e.position));
-    GoogleMapsFlutterPlatform.instance.onMarkerDrag(mapId: mapId).listen(
-        (MarkerDragEvent e) =>
-            _googleMapState.onMarkerDrag(e.value, e.position));
-    GoogleMapsFlutterPlatform.instance.onMarkerDragEnd(mapId: mapId).listen(
-        (MarkerDragEndEvent e) =>
-            _googleMapState.onMarkerDragEnd(e.value, e.position));
-    GoogleMapsFlutterPlatform.instance.onInfoWindowTap(mapId: mapId).listen(
-        (InfoWindowTapEvent e) => _googleMapState.onInfoWindowTap(e.value));
+    GoogleMapsFlutterPlatform.instance
+        .onMarkerDragStart(mapId: mapId)
+        .listen((MarkerDragStartEvent e) => _googleMapState.onMarkerDragStart(e.value, e.position));
+    GoogleMapsFlutterPlatform.instance
+        .onMarkerDrag(mapId: mapId)
+        .listen((MarkerDragEvent e) => _googleMapState.onMarkerDrag(e.value, e.position));
+    GoogleMapsFlutterPlatform.instance
+        .onMarkerDragEnd(mapId: mapId)
+        .listen((MarkerDragEndEvent e) => _googleMapState.onMarkerDragEnd(e.value, e.position));
+    GoogleMapsFlutterPlatform.instance
+        .onInfoWindowTap(mapId: mapId)
+        .listen((InfoWindowTapEvent e) => _googleMapState.onInfoWindowTap(e.value));
     GoogleMapsFlutterPlatform.instance
         .onPolylineTap(mapId: mapId)
         .listen((PolylineTapEvent e) => _googleMapState.onPolylineTap(e.value));
@@ -77,11 +79,16 @@ class GoogleMapController {
     GoogleMapsFlutterPlatform.instance
         .onTap(mapId: mapId)
         .listen((MapTapEvent e) => _googleMapState.onTap(e.position));
-    GoogleMapsFlutterPlatform.instance.onLongPress(mapId: mapId).listen(
-        (MapLongPressEvent e) => _googleMapState.onLongPress(e.position));
+    GoogleMapsFlutterPlatform.instance
+        .onLongPress(mapId: mapId)
+        .listen((MapLongPressEvent e) => _googleMapState.onLongPress(e.position));
     GoogleMapsFlutterPlatform.instance
         .onClusterTap(mapId: mapId)
         .listen((ClusterTapEvent e) => _googleMapState.onClusterTap(e.value));
+
+    GoogleMapsFlutterPlatform.instance
+        .onGroundOverlayTap(mapId: mapId)
+        .listen((GroundOverlayTapEvent e) => _googleMapState.onGroundOverlayTap(e.value));
   }
 
   /// Updates configuration options of the map user interface.
@@ -91,8 +98,7 @@ class GoogleMapController {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updateMapConfiguration(MapConfiguration update) {
-    return GoogleMapsFlutterPlatform.instance
-        .updateMapConfiguration(update, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.updateMapConfiguration(update, mapId: mapId);
   }
 
   /// Updates marker configuration.
@@ -102,8 +108,7 @@ class GoogleMapController {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updateMarkers(MarkerUpdates markerUpdates) {
-    return GoogleMapsFlutterPlatform.instance
-        .updateMarkers(markerUpdates, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.updateMarkers(markerUpdates, mapId: mapId);
   }
 
   /// Updates cluster manager configuration.
@@ -112,8 +117,7 @@ class GoogleMapController {
   /// platform side.
   ///
   /// The returned [Future] completes after listeners have been notified.
-  Future<void> _updateClusterManagers(
-      ClusterManagerUpdates clusterManagerUpdates) {
+  Future<void> _updateClusterManagers(ClusterManagerUpdates clusterManagerUpdates) {
     return GoogleMapsFlutterPlatform.instance
         .updateClusterManagers(clusterManagerUpdates, mapId: mapId);
   }
@@ -125,8 +129,7 @@ class GoogleMapController {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updatePolygons(PolygonUpdates polygonUpdates) {
-    return GoogleMapsFlutterPlatform.instance
-        .updatePolygons(polygonUpdates, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.updatePolygons(polygonUpdates, mapId: mapId);
   }
 
   /// Updates polyline configuration.
@@ -136,8 +139,7 @@ class GoogleMapController {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updatePolylines(PolylineUpdates polylineUpdates) {
-    return GoogleMapsFlutterPlatform.instance
-        .updatePolylines(polylineUpdates, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.updatePolylines(polylineUpdates, mapId: mapId);
   }
 
   /// Updates circle configuration.
@@ -147,8 +149,18 @@ class GoogleMapController {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updateCircles(CircleUpdates circleUpdates) {
+    return GoogleMapsFlutterPlatform.instance.updateCircles(circleUpdates, mapId: mapId);
+  }
+
+  /// Updates ground overlay configuration.
+  ///
+  /// Change listeners are notified once the update has been made on the
+  /// platform side.
+  ///
+  /// The returned [Future] completes after listeners have been notified.
+  Future<void> _updateGroundOverlays(GroundOverlayUpdates groundOverlayUpdates) {
     return GoogleMapsFlutterPlatform.instance
-        .updateCircles(circleUpdates, mapId: mapId);
+        .updateGroundOverlays(groundOverlayUpdates, mapId: mapId);
   }
 
   /// Updates heatmap configuration.
@@ -158,8 +170,7 @@ class GoogleMapController {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> _updateHeatmaps(HeatmapUpdates heatmapUpdates) {
-    return GoogleMapsFlutterPlatform.instance
-        .updateHeatmaps(heatmapUpdates, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.updateHeatmaps(heatmapUpdates, mapId: mapId);
   }
 
   /// Updates tile overlays configuration.
@@ -181,8 +192,7 @@ class GoogleMapController {
   /// in-memory cache of tiles. If you want to cache tiles for longer, you
   /// should implement an on-disk cache.
   Future<void> clearTileCache(TileOverlayId tileOverlayId) async {
-    return GoogleMapsFlutterPlatform.instance
-        .clearTileCache(tileOverlayId, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.clearTileCache(tileOverlayId, mapId: mapId);
   }
 
   /// Starts an animated change of the map camera position.
@@ -190,8 +200,7 @@ class GoogleMapController {
   /// The returned [Future] completes after the change has been started on the
   /// platform side.
   Future<void> animateCamera(CameraUpdate cameraUpdate) {
-    return GoogleMapsFlutterPlatform.instance
-        .animateCamera(cameraUpdate, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.animateCamera(cameraUpdate, mapId: mapId);
   }
 
   /// Changes the map camera position.
@@ -199,8 +208,7 @@ class GoogleMapController {
   /// The returned [Future] completes after the change has been made on the
   /// platform side.
   Future<void> moveCamera(CameraUpdate cameraUpdate) {
-    return GoogleMapsFlutterPlatform.instance
-        .moveCamera(cameraUpdate, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.moveCamera(cameraUpdate, mapId: mapId);
   }
 
   /// Sets the styling of the base map.
@@ -218,8 +226,7 @@ class GoogleMapController {
   /// style reference for more information regarding the supported styles.
   @Deprecated('Use GoogleMap.style instead.')
   Future<void> setMapStyle(String? mapStyle) {
-    return GoogleMapsFlutterPlatform.instance
-        .setMapStyle(mapStyle, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.setMapStyle(mapStyle, mapId: mapId);
   }
 
   /// Returns the last style error, if any.
@@ -238,8 +245,7 @@ class GoogleMapController {
   /// Screen location is in screen pixels (not display pixels) with respect to the top left corner
   /// of the map, not necessarily of the whole screen.
   Future<ScreenCoordinate> getScreenCoordinate(LatLng latLng) {
-    return GoogleMapsFlutterPlatform.instance
-        .getScreenCoordinate(latLng, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.getScreenCoordinate(latLng, mapId: mapId);
   }
 
   /// Returns [LatLng] corresponding to the [ScreenCoordinate] in the current map view.
@@ -247,8 +253,7 @@ class GoogleMapController {
   /// Returned [LatLng] corresponds to a screen location. The screen location is specified in screen
   /// pixels (not display pixels) relative to the top left of the map, not top left of the whole screen.
   Future<LatLng> getLatLng(ScreenCoordinate screenCoordinate) {
-    return GoogleMapsFlutterPlatform.instance
-        .getLatLng(screenCoordinate, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.getLatLng(screenCoordinate, mapId: mapId);
   }
 
   /// Programmatically show the Info Window for a [Marker].
@@ -260,8 +265,7 @@ class GoogleMapController {
   ///   * [hideMarkerInfoWindow] to hide the Info Window.
   ///   * [isMarkerInfoWindowShown] to check if the Info Window is showing.
   Future<void> showMarkerInfoWindow(MarkerId markerId) {
-    return GoogleMapsFlutterPlatform.instance
-        .showMarkerInfoWindow(markerId, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.showMarkerInfoWindow(markerId, mapId: mapId);
   }
 
   /// Programmatically hide the Info Window for a [Marker].
@@ -273,8 +277,7 @@ class GoogleMapController {
   ///   * [showMarkerInfoWindow] to show the Info Window.
   ///   * [isMarkerInfoWindowShown] to check if the Info Window is showing.
   Future<void> hideMarkerInfoWindow(MarkerId markerId) {
-    return GoogleMapsFlutterPlatform.instance
-        .hideMarkerInfoWindow(markerId, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.hideMarkerInfoWindow(markerId, mapId: mapId);
   }
 
   /// Returns `true` when the [InfoWindow] is showing, `false` otherwise.
@@ -286,8 +289,7 @@ class GoogleMapController {
   ///   * [showMarkerInfoWindow] to show the Info Window.
   ///   * [hideMarkerInfoWindow] to hide the Info Window.
   Future<bool> isMarkerInfoWindowShown(MarkerId markerId) {
-    return GoogleMapsFlutterPlatform.instance
-        .isMarkerInfoWindowShown(markerId, mapId: mapId);
+    return GoogleMapsFlutterPlatform.instance.isMarkerInfoWindowShown(markerId, mapId: mapId);
   }
 
   /// Returns the current zoom level of the map