+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/GoogleMapController.m
@@ -5,7 +5,7 @@
 @import GoogleMapsUtils;
 
 #import "GoogleMapController.h"
-
+#import "GoogleMapGroundOverlayController.h"
 #import "FGMMarkerUserData.h"
 #import "FLTGoogleMapHeatmapController.h"
 #import "FLTGoogleMapJSONConversions.h"